Add windows support
