+++ .github/workflows/rust.yml
@@ -59,10 +59,14 @@ jobs:
       run: cargo build --release --verbose
 
   unit-test:
-    runs-on: ubuntu-24.04
+    runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
+        os:
+          - ubuntu-24.04
+          - macos-latest
+          - windows-latest
         features:
           - name: default
           - name: none
